Add files via upload
Backup conditions that work when there are AGs in the environment. They also include user databases not in an AG, but do not include the system databases. If you have a mix of AG and non-AG environments, you can use these "with AGs" conditions and do not also need the other backup conditions without AG.